requirements.txt
Flask>=2.0
python-dotenv
psycopg2-binary
openai
gunicorn
pandas
numpy
WeasyPrint
Werkzeug
Jinja2
pillow